requirements.txt
aegea==3.2.7
pyOpenSSL==19.0.0
setuptools==41.0.1
miniwdl==0.7.4